refacto(pkh): edit /pkh routes to match /u
