+++ src/pages/pkh/[id]/creations.tsx
@@ -2,30 +2,24 @@ import { User } from '../../../types/entities/User'
 import { UserProfileLayout } from "../../../containers/User/UserProfileLayout"
 import { getServerSidePropsUserById } from "../../../services/ServerSideProps/ServerSidePropsUser"
 import { ReactElement } from "react"
-import { UserActions } from '../../../containers/User/UserActions'
-import { UserSalesTable } from "../../../containers/User/UserSalesTable";
-import { Spacing } from "../../../components/Layout/Spacing";
+import { UserGenerativeTokens } from "../../../containers/User/UserGenerativeTokens"
 
 
 interface Props {
   user: User
 }
 
-const UserPageDashboard = ({ user }: Props) => {
+const UserPageCreations = ({ user }: Props) => {
   return (
-    <>
-      <UserSalesTable user={user} />
-      <Spacing size="5x-large" />
-      <UserActions user={user}/>
-    </>
+    <UserGenerativeTokens user={user}/>
   )
 }
 
-UserPageDashboard.getLayout = function getLayout(page: ReactElement) {
+UserPageCreations.getLayout = function getLayout(page: ReactElement) {
   return (
     <UserProfileLayout
       user={page.props.user}
-      tabIndex={3}
+      tabIndex={1}
     >
       {page}
     </UserProfileLayout>
@@ -34,4 +28,4 @@ UserPageDashboard.getLayout = function getLayout(page: ReactElement) {
 
 export const getServerSideProps = getServerSidePropsUserById
 
-export default UserPageDashboard
+export default UserPageCreations
