+++ src/pages/pkh/[id]/index.tsx
@@ -2,20 +2,20 @@ import { User } from '../../../types/entities/User'
 import { UserProfileLayout } from "../../../containers/User/UserProfileLayout"
 import { getServerSidePropsUserById } from "../../../services/ServerSideProps/ServerSidePropsUser"
 import { ReactElement } from "react"
-import { UserGenerativeTokens } from "../../../containers/User/UserGenerativeTokens"
+import { UserDashboard } from "../../../containers/User/UserDashboard";
 
 
 interface Props {
   user: User
 }
 
-const UserPageCreations = ({ user }: Props) => {
+const UserPageDashboard = ({ user }: Props) => {
   return (
-    <UserGenerativeTokens user={user}/>
+    <UserDashboard user={user} />
   )
 }
 
-UserPageCreations.getLayout = function getLayout(page: ReactElement) {
+UserPageDashboard.getLayout = function getLayout(page: ReactElement) {
   return (
     <UserProfileLayout
       user={page.props.user}
@@ -28,4 +28,4 @@ UserPageCreations.getLayout = function getLayout(page: ReactElement) {
 
 export const getServerSideProps = getServerSidePropsUserById
 
-export default UserPageCreations
+export default UserPageDashboard
